sm8150-common: Drop `vendor.lineage.livedisplay`
The LiveDisplay HAL was heavly dependent on the LinageOS Settings, and on other LineageOS
HALs. It also required a lot of tweaks ROM-side to the SEPolicy etc. so it was (and still is)
a mess to implement to non-LineageOS based ROMs.

That's why I decided to switch to PanelMode, which does the same exact same thing but without
any weird LineageOS dependency. It's based on AOSiP's DeviceSettings with modification
made by [@]timocapa to have the only function to change Panel colors.